修正轨迹转动方向，转角时检测转动是否停止

+++ app/src/main/java/demo/orient/OrientCallBack.java
@@ -8,5 +8,5 @@ public interface  OrientCallBack {
     /**
      * 方向回调
      */
-    public void Orient(float orient);
+    public void Orient(int orient);
 }
